moved deletion tests to their own class
